Implemented saving information about nn model
When training is performed on a series of different configurations
of networks, we need to know which of the networks was most suc-
cessful and what was the configuration of the network. That is
why we need to save information about configuration so that we
can restore the network later.