Removed referencing dataset from arc_detect
It makes sense that arc_detect deals with the model of the
network and nn_config, but it does not makes sense that it tries
to get some data about the model from the dataset, as the dataset
could have been changed since the model was trained. Therefore
I removed referenced to the dataset.