Added visualization of CNN output
In arc_detect.py added possibility to visualize output of
the CNN, so that it is possible to understand, which features the
CNN extracts. Besides, in arc_detect also added printing of
titles of detected objects to the image, and made printing dynamic
(i.e. number of printed information depends on the number of
objects detected in an image).